% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{add,JD3_SsfEquation,character-method}
\alias{add,JD3_SsfEquation,character-method}
\title{Title}
\usage{
\S4method{add}{JD3_SsfEquation,character}(object, item, coeff = 1, fixed = TRUE, loading = NULL)
}
\arguments{
\item{object}{JD3_SsfEquation.}

\item{item}{character.}
}
\value{

}
\description{
Title
}
