% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{add,JD3_SsfModel,JD3_SsfStateBlock-method}
\alias{add,JD3_SsfModel,JD3_SsfStateBlock-method}
\title{Title}
\usage{
\S4method{add}{JD3_SsfModel,JD3_SsfStateBlock}(object, item)
}
\arguments{
\item{object}{JD3_SsfModel.}

\item{item}{JD3_SsfStateBlock.}
}
\value{

}
\description{
Title
}
