% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{ar}
\alias{ar}
\title{Title}
\usage{
ar(
  name,
  ar,
  fixedar = FALSE,
  variance = 0.01,
  fixedvariance = FALSE,
  nlags = 0,
  zeroinit = FALSE
)
}
\arguments{
\item{name}{}

\item{ar}{}

\item{fixedar}{}

\item{variance}{}

\item{fixedvariance}{}

\item{nlags}{}

\item{zeroinit}{}
}
\value{

}
\description{
Title
}
