% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{ar2}
\alias{ar2}
\title{Title}
\usage{
ar2(
  name,
  ar,
  fixedar = FALSE,
  variance = 0.01,
  fixedvariance = FALSE,
  nlags = 0,
  nfcasts = 0
)
}
\arguments{
\item{name}{}

\item{ar}{}

\item{fixedar}{}

\item{variance}{}

\item{fixedvariance}{}

\item{nlags}{}

\item{nfcasts}{}
}
\value{

}
\description{
Title
}
