% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{arima}
\alias{arima}
\title{Title}
\usage{
arima(name, ar, diff, ma, var = 1, fixed = FALSE)
}
\arguments{
\item{name}{}

\item{ar}{}

\item{diff}{}

\item{ma}{}

\item{var}{}

\item{fixed}{}
}
\value{

}
\description{
Title
}
