% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{arma}
\alias{arma}
\title{Title}
\usage{
arma(name, ar, fixedar = F, ma, fixedma = F, var = 1, fixedvar = FALSE)
}
\arguments{
\item{name}{}

\item{ar}{}

\item{fixedar}{}

\item{ma}{}

\item{fixedma}{}

\item{var}{}

\item{fixedvar}{}
}
\value{

}
\description{
Title
}
