% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_utility.R
\name{calendar}
\alias{calendar}
\title{Title}
\usage{
calendar(
  group,
  contrasts = TRUE,
  period,
  startYear,
  startPeriod = 1,
  length = 0
)
}
\arguments{
\item{group}{}

\item{contrasts}{}

\item{period}{}

\item{startYear}{}

\item{startPeriod}{}

\item{length}{}
}
\value{

}
\description{
Title
}
