% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{compute,JD3_SsfModel-method}
\alias{compute,JD3_SsfModel-method}
\title{Title}
\usage{
\S4method{compute}{JD3_SsfModel}(object, data, parameters, marginal = FALSE, concentrated = TRUE)
}
\arguments{
\item{object}{JD3_SsfModel.}
}
\value{

}
\description{
Title
}
