% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{cycle}
\alias{cycle}
\title{Title}
\usage{
cycle(
  name,
  factor = 0.9,
  period = 60,
  fixed = F,
  variance = 0.01,
  fixedvariance = FALSE
)
}
\arguments{
\item{name}{}

\item{factor}{}

\item{period}{}

\item{fixed}{}

\item{variance}{}

\item{fixedvariance}{}
}
\value{

}
\description{
Title
}
