% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procresults.R
\name{dictionary,JD3_ProcResults-method}
\alias{dictionary,JD3_ProcResults-method}
\title{Title}
\usage{
\S4method{dictionary}{JD3_ProcResults}(object)
}
\arguments{
\item{JD3_ProcResults}{}
}
\value{

}
\description{
Title
}
