% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{estimate,JD3_SsfModel-method}
\alias{estimate,JD3_SsfModel-method}
\title{Title}
\usage{
\S4method{estimate}{JD3_SsfModel}(
  object,
  data,
  marginal = F,
  concentrated = T,
  initialization = c("Diffuse", "SqrtDiffuse", "Augmented"),
  optimizer = c("LevenbergMarquardt", "MinPack", "BFGS", "LBFGS"),
  precision = 1e-15,
  initialParameters = NULL
)
}
\arguments{
\item{JD3_SsfModel}{}
}
\value{

}
\description{
Title
}
