% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procresults.R
\docType{class}
\name{JD3_Object-class}
\alias{JD3_Object-class}
\title{Title}
\value{



}
\description{
Title

Title
}
\section{Slots}{

\describe{
\item{\code{internal}}{jobjRef.}

\item{\code{internal}}{jobjRef.}
}}

