% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{loading,JD3_SsfModelEstimation-method}
\alias{loading,JD3_SsfModelEstimation-method}
\title{Title}
\usage{
\S4method{loading}{JD3_SsfModelEstimation}(object, obs = 1)
}
\arguments{
\item{object}{JD3_SsfModelEstimation.}
}
\value{

}
\description{
Title
}
