% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{locallevel}
\alias{locallevel}
\title{Title}
\usage{
locallevel(name, variance = 0.01, fixed = FALSE, initial = NaN)
}
\arguments{
\item{name}{}

\item{variance}{}

\item{fixed}{}

\item{initial}{}
}
\value{

}
\description{
Title
}
