% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{locallineartrend}
\alias{locallineartrend}
\title{Title}
\usage{
locallineartrend(
  name,
  levelVariance = 0.01,
  slopevariance = 0.01,
  fixedLevelVariance = FALSE,
  fixedSlopeVariance = FALSE
)
}
\arguments{
\item{name}{}

\item{levelVariance}{}

\item{slopevariance}{}

\item{fixedLevelVariance}{}

\item{fixedSlopeVariance}{}
}
\value{

}
\description{
Title
}
