% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{msae}
\alias{msae}
\title{Title}
\usage{
msae(name, nwaves, ar, fixedar = TRUE, lag = 1)
}
\arguments{
\item{name}{}

\item{nwaves}{}

\item{ar}{}

\item{fixedar}{}

\item{lag}{}
}
\value{

}
\description{
Title
}
