% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{msae2}
\alias{msae2}
\title{Title}
\usage{
msae2(name, vars, fixedvars = F, ar, fixedar = T, lag = 1)
}
\arguments{
\item{name}{}

\item{vars}{}

\item{fixedvars}{}

\item{ar}{}

\item{fixedar}{}

\item{lag}{}
}
\value{

}
\description{
Title
}
