% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procresults.R
\name{result,JD3_ProcResults,character-method}
\alias{result,JD3_ProcResults,character-method}
\title{Title}
\usage{
\S4method{result}{JD3_ProcResults,character}(object, id)
}
\arguments{
\item{object}{JD3_ProcResults.}

\item{id}{character.}
}
\value{

}
\description{
Title
}
