% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{sae}
\alias{sae}
\title{Title}
\usage{
sae(name, ar, fixedar = FALSE, lag = 1, zeroinit = FALSE)
}
\arguments{
\item{name}{}

\item{ar}{}

\item{fixedar}{}

\item{lag}{}

\item{zeroinit}{}
}
\value{

}
\description{
Title
}
