% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{sarima}
\alias{sarima}
\title{Title}
\usage{
sarima(
  name,
  period,
  orders,
  seasonal,
  parameters = NULL,
  fixedparameters = FALSE,
  var = 1,
  fixedvariance = FALSE
)
}
\arguments{
\item{name}{}

\item{period}{}

\item{orders}{}

\item{seasonal}{}

\item{parameters}{}

\item{fixedparameters}{}

\item{var}{}

\item{fixedvariance}{}
}
\value{

}
\description{
Title
}
