% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{seasonal}
\alias{seasonal}
\title{Title}
\usage{
seasonal(
  name,
  period,
  type = c("Trigonometric", "Crude", "HarrisonStevens", "Dummy"),
  variance = 0.01,
  fixed = FALSE
)
}
\arguments{
\item{name}{}

\item{period}{}

\item{type}{}

\item{variance}{}

\item{fixed}{}
}
\value{

}
\description{
Title
}
