% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{signal,JD3_SsfModelEstimation-method}
\alias{signal,JD3_SsfModelEstimation-method}
\title{Title}
\usage{
\S4method{signal}{JD3_SsfModelEstimation}(object, obs = 1, pos = NULL, loading = NULL, stdev = F)
}
\arguments{
\item{object}{JD3_SsfModelEstimation.}
}
\value{

}
\description{
Title
}
