% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_ssf.R
\name{weightednoise}
\alias{weightednoise}
\title{Title}
\usage{
weightednoise(name, weights, variance = 0.01, fixed = FALSE)
}
\arguments{
\item{name}{}

\item{weights}{}

\item{variance}{}

\item{fixed}{}
}
\value{

}
\description{
Title
}
